Add the characters method which takes in type, string, key, date
