+++ lib/enigma.rb
@@ -23,13 +23,7 @@ class Enigma
     {decryption: output, key: key, date: date}
   end
 
-  def generate_keys(key, date)
-    a_key = Key.new((key[0..1]), date, 3)
-    b_key = Key.new((key[1..2]), date, 2)
-    c_key = Key.new((key[2..3]), date, 1)
-    d_key = Key.new((key[3..4]), date, 0)
-    [d_key, a_key, b_key, c_key]
-  end
+  
 
   def generate_key
     rand.to_s[3..7]
