+++ test/shift_test.rb
@@ -9,6 +9,11 @@ class ShiftTest < MiniTest::Test
   def test_does_it_exist
     assert_instance_of Shift, @shift
   end
+
+  def test_it_will_return_correct_keys
+    assert_equal :decryption, @shift.characters(:decryption, 'hello', '02715', '040895').keys[0]
+    assert_equal :encryption, @shift.characters(:encryption, 'hello', '02715', '040895').keys[0]
+  end
   
 
 end
